requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
chardet==5.2.0
Django==5.2
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-qr-code==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
qrcode==8.1
reportlab==4.4.1
segno==1.6.6
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
